requirements.txt
python-dotenv
openai
pydantic
langchain